Use BlazablePost type for post items
