+++ client/my-sites/promote-post-i2/components/posts-table/index.tsx
@@ -1,11 +1,12 @@
 import '../campaigns-table/style.scss';
 
-import PostItem, { BlazablePost } from 'calypso/my-sites/promote-post-i2/components/post-item';
+import { BlazablePost } from 'calypso/data/promote-post/types';
+import PostItem from 'calypso/my-sites/promote-post-i2/components/post-item';
 import { CampaignItemLoading } from '../campaigns-table';
 import PostsListHeader from '../posts-list/header';
 
 interface Props {
-	posts: any;
+	posts: BlazablePost[];
 	isLoading: boolean;
 	isFetchingPageResults: boolean;
 }
